+++ src/typography.rs
@@ -1,15 +1,13 @@
-// #![allow(unused_variables)]
-// #![allow(unused_mut)]
-// #![allow(unused_imports)]
 #![allow(dead_code)]
 #![allow(non_snake_case)]
 use std::ops::Range;
+use std::iter::zip;
 use neon::prelude::*;
 use serde_json::{json, Value};
-use skia_safe::{Font, FontMetrics, Typeface, Paint, Point, Rect, Path as SkPath, Color};
+use skia_safe::{FontMetrics, Typeface, Paint, Point, Rect, Path as SkPath, Color};
 use skia_safe::font_style::{FontStyle, Weight, Width, Slant};
 use skia_safe::textlayout::{
-  paragraph::FontInfo, Decoration, FontCollection, Paragraph, ParagraphBuilder, ParagraphStyle, RectHeightStyle, RectWidthStyle,
+  Decoration, FontCollection, Paragraph, ParagraphBuilder, ParagraphStyle, RectHeightStyle, RectWidthStyle,
   TextAlign, TextDecoration, TextDecorationMode, TextDecorationStyle, TextDirection, TextStyle,
 };
 use crate::font_library::FontLibrary;
@@ -80,41 +78,115 @@ impl Typesetter{
 
   pub fn metrics(&self) -> Value {
     let (mut paragraph, origin) = self.layout(&Paint::default());
-    let font_ranges = paragraph.get_fonts(); // char ranges of contiguous fonts (and their metrics)
     let lookup = CharMap::new(&self.text, &paragraph); // the byte-idx -> glyph/char-idx mapping
+    let mut line_rects:Vec<Rect> = vec![]; // accumulate line rects to calculate full bounds
 
-    // calculate baseline positions (as offsets to origin based on current ctx.textBaseline setting)
+    // calculate baseline offsets (relative to line_metrics.baseline which reflects ctx.textBaseline setting)
     let shift = self.char_style.baseline_shift();
     let hang = Baseline::Hanging.get_offset(&self.char_style) - shift;
     let norm = Baseline::Alphabetic.get_offset(&self.char_style) - shift;
     let ideo = Baseline::Ideographic.get_offset(&self.char_style) - shift;
 
+    // calculate bounds for each single-font block of glyphs on each line (and gather font info)
+    struct TextRun{ line: usize, family: String, metrics: FontMetrics, bounds: Rect }
+    let mut text_runs:Vec<TextRun> = vec![];
+    paragraph.extended_visit(|line, visit|{
+      if let Some(info) = visit{
+        text_runs.push(TextRun{
+          line,
+          family: info.font().typeface().family_name(),
+          metrics: info.font().metrics().1,
+          bounds: zip(info.positions(), info.bounds())
+            .filter(|(_, rect)| !rect.is_empty())
+            .map(|(pt, rect)| rect.with_offset(*pt + info.origin() + origin - Point::new(0.0, norm)))
+            .reduce(Rect::join2)
+            .unwrap_or(Rect::new_empty())
+        });
+      }
+    });
+
+    // measure each line and add its layout rect to `line_rects`
+    let lines = (0..paragraph.line_number()).filter_map(|ln|{
+      // find the range of byte & char indices that are on this line (includes trailing whitespace if not wrapping)
+      let text_range = paragraph.get_actual_text_range(ln, !self.text_wrap);
+      let char_range = lookup.char_range(&text_range);
+      let glyph_range = lookup.glyph_range(&text_range);
+
+      // calculate this line's vertical offsets relative to the typesetting origin
+      let line_metrics = paragraph.get_line_metrics_at(ln)?;
+      let half_leading = self.graf_style.strut_style().leading().max(0.0) * self.char_style.font_size() / 2.0;
+      let baseline = line_metrics.baseline as f32 + origin.y - half_leading;
+      let line_ascent = baseline - line_metrics.ascent as f32;
+      let line_descent = baseline + line_metrics.descent as f32;
+
+      // combine the glyph bounds of all single-font runs on this line (potentially omitting trailing spaces)
+      let font_runs = text_runs.iter().filter(|r| r.line==ln).collect::<Vec<&TextRun>>();
+      let text_bounds = font_runs.iter()
+        .map(|run| run.bounds)
+        .reduce(Rect::join2)
+        .unwrap_or(Rect::new_empty());
+
+      // calculate horizontal line bounds that include trailing whitespace for use in `actualBoundingBox`
+      // (and compensate for the extra half-letterspace added to the start & end of each line)
+      line_rects.push(
+        paragraph
+          .get_rects_for_range(glyph_range, RectHeightStyle::Tight, RectWidthStyle::Tight).iter()
+          .map(|tb| {
+            let Rect{top, bottom, ..} = text_bounds;
+            let Rect{left, right, ..} = tb.rect.with_offset(origin);
+            Rect::new(left, top, right - self.char_style.letter_spacing(), bottom)
+          })
+          .reduce(Rect::join2)
+          .unwrap_or(Rect::new_empty())
+      );
+
+      Some(json!({
+        "x": text_bounds.left,
+        "y": text_bounds.top,
+        "width": text_bounds.width(),
+        "height": text_bounds.height(),
+        "baseline": baseline, // corresponds to the ctx.textBaseline selection
+        "hangingBaseline": baseline - hang,
+        "alphabeticBaseline": baseline - norm,
+        "ideographicBaseline": baseline - ideo,
+        "ascent": line_ascent,
+        "descent": line_descent,
+        "startIndex": char_range.start,
+        "endIndex": char_range.end,
+        "runs": font_runs.iter().map(|TextRun{family, metrics, bounds, ..}| {
+          json!({
+            "x": bounds.left,
+            "y": bounds.top,
+            "width": bounds.width(),
+            "height": bounds.height(),
+            "family": family,
+            "ascent": baseline - norm + metrics.ascent,
+            "descent": baseline - norm + metrics.descent,
+            "capHeight": baseline - norm - metrics.cap_height,
+            "xHeight": baseline - norm - metrics.x_height,
+            "underline": metrics.underline_position().map(|ulH| baseline - norm + ulH ),
+            "strikethrough": metrics.strikeout_position().map(|stH| baseline - norm + stH ),
+          })
+        }).collect::<Vec<Value>>()
+      }))
+    }).collect::<Vec<Value>>();
+
+    // combine all the individual line measurements to find the `actualBoundingBox`
+    let full_bounds = line_rects.into_iter()
+      .reduce(Rect::join2)
+      .unwrap_or(Rect::new_empty());
+
     // use line metrics to find maximal ascent/descent of all fonts on first line
     let (ascent, descent) = paragraph.get_line_metrics_at(0).map(|line|
       (norm + line.ascent as f32, line.descent as f32 - norm)
     ).unwrap_or_else(||{
-      // or fall back to the first-matched font's metrics
+      // or fall back to the first-matched font's metrics if measuring empty string
       let FontMetrics{ascent, descent, ..} = self.char_style.font_metrics();
       (norm - ascent, descent - norm)
     });
 
-    // helper to line up layout rects with the origin and compensate for letterspacing
-    let adjust_bounds = |rect:&Rect| -> Rect{
-      let Rect{left, top, right, bottom} = *rect;
-      Rect::new(left, top, right - self.char_style.letter_spacing(), bottom).with_offset(origin)
-    };
-
-    // take the union of each line's bounds to construct the whole-text-run bounds
-    let full_bounds:Rect = (0..paragraph.line_number())
-      .flat_map(|ln| paragraph.get_rects_for_range(
-        lookup.glyph_range(&paragraph.get_actual_text_range(ln, !self.text_wrap)), RectHeightStyle::Tight, RectWidthStyle::Tight
-      ))
-      .map(|tb| adjust_bounds(&tb.rect))
-      .reduce(Rect::join2)
-      .unwrap_or(Rect::new_empty());
-
     json!({
-      "width": (-full_bounds.left).max(0.0) + full_bounds.right,
+      "width": full_bounds.right - full_bounds.left,
       "actualBoundingBoxLeft": -full_bounds.left,
       "actualBoundingBoxRight": full_bounds.right,
       "actualBoundingBoxAscent": -full_bounds.top,
@@ -126,83 +198,7 @@ impl Typesetter{
       "hangingBaseline": hang,
       "alphabeticBaseline": norm,
       "ideographicBaseline": ideo,
-      "lines": (0..paragraph.line_number()).filter_map(|ln|{
-        // find the range of byte & char indices that are on this line (includes trailing whitespace if not wrapping)
-        let text_range = paragraph.get_actual_text_range(ln, !self.text_wrap);
-        let char_range = lookup.char_range(&text_range);
-
-        // calculate this line's vertical offset relative to the typesetting origin
-        let line_metrics = paragraph.get_line_metrics_at(ln)?;
-        let half_leading = self.graf_style.strut_style().leading().max(0.0) * self.char_style.font_size() / 2.0;
-        let baseline = line_metrics.baseline as f32 + origin.y - half_leading;
-        let ascent = line_metrics.ascent as f32;
-        let descent = line_metrics.descent as f32;
-
-        // divide line into runs of contiguous font use
-        let font_runs:Vec<(&Font, Range<usize>)> = font_ranges.iter()
-          .filter_map(|FontInfo{text_range: font_range, font}|{
-            match font_range.start.max(text_range.start)..font_range.end.min(text_range.end){
-              rng if !rng.is_empty() => Some((font, rng)),
-              _ => None
-            }
-          })
-          .collect();
-
-        // find the full-line bounds (either tight or loose depending on number of fonts present)
-        let line_bounds = if font_runs.len() == 1 {
-          // use path-tracing to get tighter bounds (but only if no font-mixing occurs)
-          // NB: avoids a Skia bug where incorrect paths are produced when outlining multi-font text runs
-          let headroom = shift + paragraph.alphabetic_baseline() - ascent;
-          adjust_bounds(
-            &paragraph.get_path_at(ln).1.compute_tight_bounds().with_offset((0.0, -headroom))
-          )
-        }else{
-          // find layout sub-rectangles (running from ascent to descent) and union them
-          paragraph
-            .get_rects_for_range(lookup.glyph_range(&text_range), RectHeightStyle::Tight, RectWidthStyle::Tight).iter()
-            .map(|tb| adjust_bounds(&tb.rect))
-            .reduce(Rect::join2)
-            .unwrap_or(Rect::new_empty())
-        };
-
-        Some(json!({
-          "x": line_bounds.left,
-          "y": line_bounds.top,
-          "width": line_bounds.width(),
-          "height": line_bounds.height(),
-          "baseline": baseline, // corresponds to the ctx.textBaseline selection
-          "hangingBaseline": baseline - hang,
-          "alphabeticBaseline": baseline - norm,
-          "ideographicBaseline": baseline - ideo,
-          "ascent": baseline - ascent,
-          "descent": baseline + descent,
-          "startIndex": char_range.start,
-          "endIndex": char_range.end,
-          "runs": font_runs.iter().filter_map(|(font, overlap)| {
-            let metrics = font.metrics().1;
-            let glyph_range = lookup.glyph_range(&overlap);
-            let base = baseline - norm;
-
-            paragraph
-              .get_rects_for_range(glyph_range, RectHeightStyle::Tight, RectWidthStyle::Tight)
-              .first() // there should only ever be one rect within a single font run
-              .map(|text_box|{
-                let rect = adjust_bounds(&text_box.rect);
-                json!({
-                  "x": rect.left,
-                  "y": rect.top,
-                  "width": rect.width(),
-                  "height": rect.height(),
-                  "family": font.typeface().family_name(),
-                  "capHeight": base - metrics.cap_height,
-                  "xHeight": base - metrics.x_height,
-                  "underline": metrics.underline_position().map(|ulH| base + ulH ),
-                  "strikethrough": metrics.strikeout_position().map(|stH| base + stH ),
-                })
-              })
-          }).collect::<Vec<Value>>()
-        }))
-      }).collect::<Vec<Value>>()
+      "lines": lines,
     })
   }